requirements.txt
cherrypy==17.3.0
celery==4.1.1
python-dateutil==2.6.0
psycopg2==2.7.3.2
py3k-bcrypt==0.3
stripe==1.25.0
pytz==2017.2
alembic==0.9.1
treepoem==1.0.1
email_validator==1.0.2
phonenumbers==8.8.1
pockets==0.6.2
residue==0.2.8
XlsxWriter==1.0.2
uszipcode==0.1.3
geopy==1.11.0
twilio==6.10.0
cherrys==0.4
redis==2.10.6
ics==0.4
pillow==6.2.0
fpdf==1.7.2
